Removed the DEFAULT_SERVER_RECEPTIONIST_PROXY_PORT constant and modified the codebase to more consistently use uint16 for port variables, rather than int32

+++ SpatialGDK/Source/SpatialGDKEditor/Public/SpatialGDKEditorSettings.h
@@ -457,7 +457,7 @@ public:
 	/** Port on which the receptionist proxy will be available. */
 	UPROPERTY(EditAnywhere, config, Category = "Cloud Connection",
 			  meta = (EditCondition = "bConnectServerToCloud", DisplayName = "Local Receptionist Port"))
-	int32 LocalReceptionistPort;
+	uint16 LocalReceptionistPort;
 
 	/** Network address to bind the receptionist proxy to. */
 	UPROPERTY(EditAnywhere, config, Category = "Cloud Connection",
