+++ SpatialGDK/Source/SpatialGDKServices/Public/LocalReceptionistProxyServerManager.h
@@ -11,12 +11,12 @@ class FLocalReceptionistProxyServerManager
 public:
 	FLocalReceptionistProxyServerManager();
 
-	bool CheckIfPortIsBound(int32 Port, FString& OutPID, FString& OutLogMessages);
-	bool LocalReceptionistProxyServerPreRunChecks(int32 ReceptionistPort);
+	bool CheckIfPortIsBound(uint16_t Port, FString& OutPID, FString& OutLogMessages);
+	bool LocalReceptionistProxyServerPreRunChecks(uint16_t ReceptionistPort);
 
-	void SPATIALGDKSERVICES_API Init(int32 ReceptionistPort);
+	void SPATIALGDKSERVICES_API Init(uint16_t ReceptionistPort);
 	bool SPATIALGDKSERVICES_API TryStartReceptionistProxyServer(bool bIsRunningInChina, const FString& CloudDeploymentName,
-																const FString& ListeningAddress, int32 ReceptionistPort);
+																const FString& ListeningAddress, uint16_t ReceptionistPort);
 	bool SPATIALGDKSERVICES_API TryStopReceptionistProxyServer();
 
 private:
@@ -27,7 +27,7 @@ private:
 
 	FProcHandle ProxyServerProcHandle;
 	FString RunningCloudDeploymentName;
-	int32 RunningProxyReceptionistPort;
+	uint16_t RunningProxyReceptionistPort;
 	FString RunningProxyListeningAddress;
 
 	bool bProxyIsRunning = false;
